Add `base-standard-library-release-notify` CI workflow.
This workflow is responsible for sending out notifications to various feeds
whenever a new release of the library is published.

The workflow is triggered on each published release in GitHub Releases.

The first supported notification is the `zulip` job, which will
publish a message in the notifications stream of the Savi instance of
the Zulip chat service using the given bot credentials.