Add `library-deps-update` CI workflow.
This workflow is responsible for helping a library repository keep
its manifest(s) continuously up to date with the latest Savi dependencies.

If the manifest(s) are found to be changed by the update, a pull request
will be opened with the changes, allowing the maintainer to pull them in.

The workflow is triggered daily, and can also be manually triggered.