Add support for "far" pointers to other segments.
The Cap'n Proto encoding allows a pointer-based entity residing
in different segment to be accessed via a "far" pointer
(or a "double-far" pointer) that allows the extra information
about what segment it resides in to be encoded.

Hence, all pointer `_parse!` methods have been updated to
first check if the pointer in question is a "far" pointer
by trying to use `CapnProto.Pointer.Far._parse!` on it,
which handles both "far" and "double-far" pointers,
then resume the normal specialized pointer logic on the
relevant values that were revised by the far pointer's logic.