Add `CapnProto.Segments.Reader` for reading from a byte stream.
This includes the logic needed to read the framing header data at
the start of the stream to determine sizes, to prepare the incoming
buffers to read the correct amount, and to actually read in the
segment data into a `CapnProto.Segments` object in a zero-copy way.

It also includes a mechanism for limiting to total size per message,
to avoid allowing an attacker to cause the reader to over-allocate
and crash due to an out of memory error. This limit is intended
to be configurable once Savi has the compile-time mechanism in
place for doing so (see https://github.com/savi-lang/savi/issues/161).