Add `library-release` CI workflow for automatically releasing.
This workflow is responsible for automating the creation of a new release.
As an input, the workflow takes a version tag to apply to the latest commit.
The version tag name will also be the release's name in GitHub.
Release notes will auto-generated by GitHub based on Pull Requests history.

With some minor modification (see comments in the job definition below),
this workflow can also build release binaries for all supported platforms,
and package them up as tarballs attached as assets to the release in GitHub.
This is not relevant for all libraries, because many libraries do not have
any executable binary that is relevant to build, so it is disabled by default,
but for those libraries that have one or more application binaries to build,
all you need to do is specify the binary manifest's name and enable that step.

The workflow is triggered by workflow dispatch, which means to run it you
need to press the "Run Workflow" button on the Actions page for this workflow,
then enter the required inputs when prompted to do so.
It can also be triggered via the GitHub API if desired.