Fix `CapnProto.List.[]!` to be partial.
Prior to this commit, accessing by index in a `CapnProto.List`
would return an empty struct instance pointing to nothing.

This was technically safe (the empty struct couldn't be used
to violate any other data, and was basically ineffective
at doing anything), but it was confusing and thus harmful.
Particularly in the `Builder` context, it was bad for
someone to think they had gotten a real struct instance
to start building data into, while it was in reality
silently discarding all their writes and doing nothing.

Now, this method will correctly raise an `error!` in
that situation instead, letting the caller know of the issue.