Fix `[]` element access on lists to take type `USize` instead of `U32`.
`USize` matches the `Indexable` trait, and is commonly used for
collections. `U32` is just an internal CapnProto implementation detail.

I wanted to go ahead and try to use the `Indexable` trait itself here,
but it's not ready for that until the language has cap specialization
features that would let us implement `CapnProto.List.Builder` methods
as `:fun` instead of `:fun ref`.