Fix union init methods for the case of primitive values and lists.
Prior to this commit there was no way to actually set the value
of a union field that led to a primitive value, and there was no
way to initialize a non-zero count for a union field that led to a list.

This commit adds those missing features by taking a paramter in
the union field init method.