Fix conversion of floating-point values for encoding.
Prior to this commit, the internal conversion process was using
value-based conversion, which tried to convert from and to the
same numeric value as the internal integers used to encode.

Instead, we need to encode using integers that correspond to
the bitwise representation of the floating-point values,
so that we have a lossless conversion to and from.