Add `trim_` methods for `CapnProto.List` fields.
This allows truncating the front or back of a list field,
freeing some memory from the head and/or tail, or freeing
the entire memory for the list if the new list count is zero.

There is no copying or moving of memory - list data stays at the same
address in the segment and it just leaves holes at the head or tail
suitable for future allocations that may fit in that space.