Fix `TraceData` for union-marked field to trace even if not set.
Prior to this commit, all fields were traced with the `if_set!`
designation to avoid tracing data which hadn't been set to
a value (and wasn't set to a non-default value).

However, for union-marked fields we want the trace to only be
conditional on whether the union was marked as such.
That is, we want to trace such fields even when their value
is set to the default value. Without tracing them we wouldn't
be tracing any indication of the union's state.