Add `capn_proto_address` to pointers and generated wrapping code.
This address is based on where the pointer is located within the
message (deterministically from the position in the list of segments).
That is, every time a given message is loaded, the `capn_proto_address`
for a given pointer within it will be the same. And the addresses
are unique (no two different pointers share the same address,
even if they are at the same offset in different segments).

Therefore, this address can be used to uniquely identify parts
of a message in a stable way across different loads of the same message.

We will use this in the Savi self-hosted compiler to track
addresses of AST nodes, within analysis dumps that refer to those nodes.