Fix issue with cross-package private access in prior release.
The last release generated code that would reach into
`CapnProto.List(T)._p`, which gives a compiler error when
it's across package boundaries.

It turns out it was unnecessary anyway, as `CapnProto.List`
already has a public method which gives what we need.

This commit fixes the code generator to no longer generate
the private accesses.